refactor: move MCP client config from dedicated page into Settings
